+++ src/components/Board-View/Community.js
