+++ src/components/img/second-logo.png
